requirements.txt
Flask==2.3.3
pymongo==4.5.0
numpy==1.24.3
scipy==1.11.3
geopy==2.4.0
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.3